+++ src/pages/Auth/Auth.js
